Создан модуль api.js с запросами к серверу
